nexuslink/pyproject.toml
[project]
name = "nexuslink"
version = "0.1.0"
description = "Cross-domain research hypothesis engine"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "networkx",
    "sentence-transformers",
    "chromadb",
    "pymupdf4llm",
    "arxiv",
    "habanero",
    "spacy",
    "fastapi",
    "uvicorn",
    "anthropic",
    "litellm",
    "bibtexparser",
    "httpx",
    "polars",
    "obsidiantools",
    "loguru",
    "pyyaml",
    "numpy",
    "jinja2",
    "click",
    "pydantic-settings",
    "python-multipart",
    "uvicorn[standard]",
    "rich",
]

[project.scripts]
nexuslink = "nexuslink.cli:cli"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# pyproject.toml lives inside nexuslink/ which itself is the importable package.
# The parent directory (Nexus-link/) is the repo root.
# We use `only-include` to select specific subdirs that Hatchling can resolve.
only-include = [
    "raw",
    "wiki",
    "llm",
    "api",
    "utils",
    "ui",
    "demo",
    "cli.py",
    "main.py",
    "config.py",
    "__init__.py",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
